Refactor settings screens and improve icon usage
- Updated various settings screens to use `rememberLazyListState` for scroll state management instead of `rememberSaveable`.
- Changed icons to use `Icons.AutoMirrored` for better alignment and appearance in settings items.
- Simplified font handling in `FontScreen` and `FontScreenViewModel` by removing unnecessary null checks.
- Enhanced notification settings with improved icon usage and layout adjustments.
- Refactored security settings to improve clarity and maintainability.
- Updated translation engine configuration to streamline locale parsing.
- Improved error handling in `BookCoverFetcher` and related image loading components.
- General code cleanup and optimization across multiple files for better readability and performance.

+++ presentation/src/commonMain/kotlin/ireader/presentation/ui/readinghub/ReadingHubViewModel.kt
@@ -137,8 +137,9 @@ class ReadingHubViewModel(
                     val currentValue = when (milestone.metric) {
                         MilestoneMetric.BOOKS_READ -> stats.booksCompleted.toLong()
                         MilestoneMetric.CHAPTERS_READ -> stats.totalChaptersRead.toLong()
-                        MilestoneMetric.READING_MINUTES -> stats.totalReadingTimeMinutes.toLong()
+                        MilestoneMetric.READING_MINUTES -> stats.totalReadingTimeMinutes
                         MilestoneMetric.STREAK_DAYS -> stats.readingStreak.toLong()
+
                     }
                     if (currentValue >= milestone.threshold) {
                         newMilestones.add(milestone)
